greengrass-components/applications/docker_button/button_app/requirements.txt
awsiotsdk==1.21.0
RPi.GPIO==0.7.1